don't switch to full precision

+++ src/transformers/models/llama/sam_embed.py
@@ -64,7 +64,7 @@ class PositionEmbeddingRandom(nn.Module):
         coords = coords_input.clone()
         coords[:, :, 0] = coords[:, :, 0] / image_size[1]
         coords[:, :, 1] = coords[:, :, 1] / image_size[0]
-        return self._pe_encoding(coords.to(torch.float))  # B x N x C
+        return self._pe_encoding(coords)  # B x N x C
 
 
     def apply_rotary_2d_pos_emb(self,q,k,coords):
